Make compatible with swell

+++ src/jedi_bundle/configure_jedi_bundle.py
@@ -30,6 +30,7 @@ def configure_jedi(logger, configure_config):
     cmake_build_type = config_get(logger, configure_config, 'cmake_build_type', 'release')
     path_to_build = config_get(logger, configure_config, 'path_to_build')
     custom_configure_options = config_get(logger, configure_config, 'custom_configure_options', '')
+    external_modules = config_get(logger, configure_config, 'external_modules', False)
 
     # Create build directory
     os.makedirs(path_to_build, exist_ok=True)
@@ -41,25 +42,39 @@ def configure_jedi(logger, configure_config):
         platform_pathfile = os.path.join(return_config_path(), 'platforms', platform + '.yaml')
         platform_dict = load_yaml(logger, platform_pathfile)
 
-        # Steps to load the chosen modules
-        modules_dict = platform_dict['modules'][modules]
-        module_inits = config_get(logger, modules_dict, 'init', [])
-        module_directives = config_get(logger, modules_dict, 'load')
-        platform_configure_directives = config_get(logger, modules_dict, 'configure', '')
+        if not external_modules:
 
-        # Create modules init file
-        modules_init = os.path.join(path_to_build, 'modules-init')
-        remove_file(logger, modules_init)
-        with open(modules_init, 'a') as modules_init_open:
-            for module_init in module_inits:
-                modules_init_open.write(module_init + '\n')
+            # Extra configure options for running tests
+            modules_dict = platform_dict['modules'][modules]
+            platform_configure_directives = config_get(logger, modules_dict, 'configure', '')
 
-        # Create modules file
-        modules_file = os.path.join(path_to_build, 'modules')
-        remove_file(logger, modules_file)
-        with open(modules_file, 'a') as modules_file_open:
-            for module_directive in module_directives:
-                modules_file_open.write(module_directive + '\n')
+            # Create modules init file
+            module_inits = config_get(logger, modules_dict, 'init', [])
+            modules_init = os.path.join(path_to_build, 'modules-init')
+            remove_file(logger, modules_init)
+            with open(modules_init, 'a') as modules_init_open:
+                for module_init in module_inits:
+                    modules_init_open.write(module_init + '\n')
+
+            # Create modules file
+            module_directives = config_get(logger, modules_dict, 'load')
+            modules_file = os.path.join(path_to_build, 'modules')
+            remove_file(logger, modules_file)
+            with open(modules_file, 'a') as modules_file_open:
+                for module_directive in module_directives:
+                    modules_file_open.write(module_directive + '\n')
+
+        # Link in CRTM coefficients to source directory to avoid download
+        crtm_coefficient_path_platform = platform_dict['crtm_coeffs_path']
+        crtm_coefficient_version = platform_dict['crtm_coeffs_version']
+        crtm_coefficient_path_source = os.path.join(path_to_source, 'test-data-release', 'crtm')
+        crtm_coefficient_target = os.path.join(crtm_coefficient_path_source,
+                                               crtm_coefficient_version)
+        os.makedirs(crtm_coefficient_path_source, exist_ok=True)
+
+        if not os.path.exists(crtm_coefficient_target):
+            os.symlink(os.path.join(crtm_coefficient_path_platform, crtm_coefficient_version), \
+                       crtm_coefficient_target)
 
     # File to hold configure steps
     configure_file = os.path.join(path_to_build, 'jedi_bundle_configure.sh')
@@ -75,7 +90,7 @@ def configure_jedi(logger, configure_config):
     with open(configure_file, 'a') as configure_file_open:
         configure_file_open.write(f'#!/usr/bin/env bash \n')
         configure_file_open.write(f'\n')
-        if platform != 'none':
+        if not external_modules:
             configure_file_open.write(f'source {modules_init}\n')
             configure_file_open.write(f'source {modules_file}\n')
         configure_file_open.write(f'\n')
